requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
undetected-chromedriver>=3.5.5
selenium>=4.18.0
setuptools>=70.0.0
pystray>=0.19.5
pillow>=10.0.0
